backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-multipart
python-dotenv
bcrypt
requests
google-generativeai
cryptography
Pillow
gunicorn
